Complete refreeze, which fixes missing images from R projects in gallery/
A previous freeze did not commit all images to version control. It is unclear which, but sufficient effort might determine this if relevant. The resulting broken freeze meant that a number of R projects were considered 'frozen' even though their images were unsaved, and were not re-run as their content was not changed. On render, the HTML then linked to nonexistant images.