Automation bugfix: ignore venv properly
a69d5fcd259900523f34dfd6e1bc69cf18ce8e2f attempts to remove the project_processing automation's requirement on venv if it's in a github actions environment (the action creates the system env `IGNORE_VENV_REQ` which Python looks for). However, the "Not in the correct virtual environment" flag, which ensures that the venv is one dir below _quarto.yml, was still going off. This commit fixes that by only running that check if  IGNORE_VENV_REQ != "true".